scrapegoat/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database & Queue
redis==5.0.1
psycopg2-binary==2.9.9

# HTTP Clients
requests==2.31.0
httpx==0.27.0
curl_cffi==0.5.10

# Resilience & Logging
tenacity==8.2.3
loguru==0.7.2

# AI & Automation
openai==1.3.0
playwright==1.40.0

# HTML Parsing (for spiders)
beautifulsoup4==4.12.3
lxml==5.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# CAPTCHA Solving (uses direct API calls, no library needed)